Norminette fixing, acabar con negativo check_float

+++ inc/struct_minirt.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   struct_minirt.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dkurcbar <dkurcbar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: iassambe <iassambe@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 19:31:29 by iassambe          #+#    #+#             */
-/*   Updated: 2024/04/29 14:52:26 by dkurcbar         ###   ########.fr       */
+/*   Updated: 2024/04/29 16:33:42 by iassambe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ typedef struct s_mlx
 
 typedef struct s_vector
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vector;
 
 typedef	struct s_rgb
@@ -43,7 +43,7 @@ typedef	struct s_rgb
 typedef struct s_sphere
 {
 	t_vector		sp_center;
-	float			sp_diam;
+	double			sp_diam;
 	t_rgb			sp_color;
 	struct s_sphere	*next;
 }	t_sphere;
@@ -60,8 +60,8 @@ typedef struct s_cylinder
 {
 	t_vector			cy_center;
 	t_vector			cy_axis;
-	float				cy_diam;
-	float				cy_height;
+	double				cy_diam;
+	double				cy_height;
 	t_rgb				cy_color;
 	struct s_cylinder	*next;
 	
@@ -76,13 +76,13 @@ typedef struct s_cylinder
 // cy = cylinder
 typedef struct s_scene
 {
-	float		a_l_ratio;
+	double		a_l_ratio;
 	t_rgb		a_l_color;
 	t_vector	c_pos;
 	t_vector	c_dir;
-	float		c_fov;
+	double		c_fov;
 	t_vector	l_pos;
-	float		l_bright;
+	double		l_bright;
 	t_sphere	*sp;
 	t_plane		*pl;
 	t_cylinder	*cy;
